requirements.txt
ipykernel
ipython
matplotlib
notebook
numpy
pandas
Pillow
seaborn
torch
torchvision
tqdm
pyiqa
blobfile
lpips
ete3
scikit-learn
scipy

